AI-Travel-Planner-Backend/requirement.txt
fastapi
uvicorn
transformers
torch
huggingface_hub
google-search-results>=2.4.2
gradio>=4.31.0
python-dotenv
